feat: Implement g+ authorization

+++ application.py
@@ -1,8 +1,17 @@
-from flask import Flask, render_template, request, redirect, url_for
+from flask import Flask, render_template, request, redirect, url_for, flash
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.exc import IntegrityError
 from database_setup import Base, Category, Item
+from flask import session as login_session
+import random
+import string
+from oauth2client.client import flow_from_clientsecrets
+from oauth2client.client import FlowExchangeError
+import httplib2
+import json
+from flask import make_response
+import requests
 
 app = Flask(__name__)
 
@@ -11,6 +20,10 @@ Base.metadata.bind = engine
 
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Catalog Application"
 
 
 def category(category_name):
@@ -38,6 +51,136 @@ def items(count='all', category_name=None):
     else:
         # return all items
         return session.query(Item).order_by('name')
+
+
+# Create anti-forgery state token
+@app.route('/login')
+def showLogin():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(
+            json.dumps('Current user is already connected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += """
+        " style = "width: 300px; height: 300px;border-radius: 150px;
+         -webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+         """
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session['access_token']
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: '
+    print login_session['username']
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = ('https://accounts.google.com/o/oauth2/revoke?token=%s'
+           % login_session['access_token'])
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+
+        response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
 
 
 @app.route('/')
@@ -72,6 +215,9 @@ def newCategory():
     """
     Allow logged users to create new category
     """
+    # if not logged in, redirect to home
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         new_category_name = request.form['name'].strip().lower()
         if new_category_name:
@@ -109,6 +255,9 @@ def editCategory(category_name):
     """
     Allow logged users to edit a category
     """
+    # if not logged in, redirect to home
+    if 'username' not in login_session:
+        return redirect('/login')
     category_to_edit = category(category_name)
     if request.method == 'POST':
         edited_category_name = request.form['name'].strip().lower()
@@ -152,6 +301,9 @@ def deleteCategory(category_name):
     """
     Allow logged users to delete a category (and items in it)
     """
+    # if not logged in, redirect to home
+    if 'username' not in login_session:
+        return redirect('/login')
     category_to_delete = category(category_name)
     items_to_delete = items(category_name=category_name)
     if request.method == 'POST':
@@ -186,6 +338,9 @@ def newItem(category_name):
     """
     Allow logged users to create an item
     """
+    # if not logged in, redirect to home
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         new_item_name = request.form['name'].strip().lower()
         new_item_description = request.form['description'].strip()
@@ -246,6 +401,9 @@ def editItem(category_name, item_name):
     """
     Allow logged users to edit an item
     """
+    # if not logged in, redirect to home
+    if 'username' not in login_session:
+        return redirect('/login')
     item_to_edit = item(item_name, category_name)
     if request.method == 'POST':
         edited_item_name = request.form['name'].strip().lower()
@@ -307,6 +465,9 @@ def deleteItem(category_name, item_name):
     """
     Allow logged users to delete an item
     """
+    # if not logged in, redirect to home
+    if 'username' not in login_session:
+        return redirect('/login')
     item_to_delete = item(item_name, category_name)
     if request.method == 'POST':
         session.delete(item_to_delete)
@@ -325,5 +486,6 @@ def deleteItem(category_name, item_name):
             item_name=item_name)
 
 if __name__ == '__main__':
+    app.secret_key = 'iamsosecret'
     app.debug = True
     app.run(host='0.0.0.0', port=5000)
